third_party/WebKit/LayoutTests/platform/linux/virtual/outofblink-cors/external/wpt/fetch/api/request/request-cache-no-cache-expected.txt

This is a testharness.js-based test.
PASS RequestCache "no-cache" mode revalidates stale responses found in the cache with Etag and stale response
PASS RequestCache "no-cache" mode revalidates stale responses found in the cache with Last-Modified and stale response
FAIL RequestCache "no-cache" mode revalidates fresh responses found in the cache with Etag and fresh response assert_equals: expected 2 but got 1
FAIL RequestCache "no-cache" mode revalidates fresh responses found in the cache with Last-Modified and fresh response assert_equals: expected 2 but got 1
Harness: the test ran to completion.
